+++ src/auth_service/claims_repository/core/seed.py
@@ -166,7 +166,7 @@ async def seed_data_steward_claims(
     This function removes all existing data steward claims and then adds such
     claims for all data stewards specified via external user ID in the config.
     """
-    if "claims" not in config.provide_apis:
+    if set(config.provide_apis).isdisjoint({"claims", "access"}):
         return
     data_stewards = config.add_as_data_stewards
     if not data_stewards:
