+++ tests/integration/claims_repository/fixtures.py
@@ -83,7 +83,7 @@ async def fixture_full_client(
         kafka_servers=kafka.config.kafka_servers,
         service_name=kafka.config.service_name,
         service_instance_id=kafka.config.service_instance_id,
-        provide_apis=["claims"],
+        provide_apis=["claims", "access"],
         add_as_data_stewards=add_as_data_stewards,
     )
     await seed_database(config)
